refactor: creates docker package.
