+++ core/core.go
@@ -5,18 +5,17 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
 	"time"
 
+	"github.com/containous/structor/docker"
 	"github.com/containous/structor/file"
 	"github.com/containous/structor/gh"
 	"github.com/containous/structor/manifest"
 	"github.com/containous/structor/menu"
 	"github.com/containous/structor/repository"
+	"github.com/containous/structor/requirements"
 	"github.com/containous/structor/types"
 	"github.com/ldez/go-git-cmd-wrapper/git"
 	"github.com/ldez/go-git-cmd-wrapper/worktree"
@@ -27,13 +26,6 @@ const (
 	baseRemote      = "origin/"
 	envVarLatestTag = "STRUCTOR_LATEST_TAG"
 )
-
-type dockerfileInformation struct {
-	name      string
-	path      string
-	content   []byte
-	imageName string
-}
 
 // Execute core process
 func Execute(config *types.Configuration) error {
@@ -60,10 +52,10 @@ func Execute(config *types.Configuration) error {
 		return errors.Wrap(err, "failed to download Dockerfile")
 	}
 
-	fallbackDockerfile := dockerfileInformation{
-		name:      fmt.Sprintf("%v.Dockerfile", time.Now().UnixNano()),
-		content:   fallbackDockerFileContent,
-		imageName: config.DockerImageName,
+	fallbackDockerfile := docker.DockerfileInformation{
+		Name:      fmt.Sprintf("%v.Dockerfile", time.Now().UnixNano()),
+		Content:   fallbackDockerFileContent,
+		ImageName: config.DockerImageName,
 	}
 
 	repoID := types.RepoID{
@@ -71,7 +63,7 @@ func Execute(config *types.Configuration) error {
 		RepositoryName: config.RepositoryName,
 	}
 
-	requirementsContent, err := getRequirementsContent(config.RequirementsURL)
+	requirementsContent, err := requirements.GetContent(config.RequirementsURL)
 	if err != nil {
 		return err
 	}
@@ -79,7 +71,7 @@ func Execute(config *types.Configuration) error {
 	return process(workDir, repoID, fallbackDockerfile, menuContent, requirementsContent, config)
 }
 
-func process(workDir string, repoID types.RepoID, fallbackDockerfile dockerfileInformation,
+func process(workDir string, repoID types.RepoID, fallbackDockerfile docker.DockerfileInformation,
 	menuContent types.MenuContent, requirementsContent []byte, config *types.Configuration) error {
 	latestTagName, err := getLatestReleaseTagName(repoID)
 	if err != nil {
@@ -114,7 +106,7 @@ func process(workDir string, repoID types.RepoID, fallbackDockerfile dockerfileI
 			return err
 		}
 
-		err = checkRequirements(versionDocsRoot)
+		err = requirements.Check(versionDocsRoot)
 		if err != nil {
 			return err
 		}
@@ -125,7 +117,7 @@ func process(workDir string, repoID types.RepoID, fallbackDockerfile dockerfileI
 			Experimental: config.ExperimentalBranchName,
 			CurrentPath:  versionDocsRoot,
 		}
-		fallbackDockerfile.path = filepath.Join(versionsInfo.CurrentPath, fallbackDockerfile.name)
+		fallbackDockerfile.Path = filepath.Join(versionsInfo.CurrentPath, fallbackDockerfile.Name)
 
 		err = buildDocumentation(branches, versionsInfo, fallbackDockerfile, menuContent, requirementsContent, config)
 		if err != nil {
@@ -177,7 +169,7 @@ func copyVersionSiteToOutputSite(versionsInfo types.VersionsInformation, siteDir
 }
 
 func buildDocumentation(branches []string, versionsInfo types.VersionsInformation,
-	fallbackDockerfile dockerfileInformation, menuTemplateContent types.MenuContent, requirementsContent []byte,
+	fallbackDockerfile docker.DockerfileInformation, menuTemplateContent types.MenuContent, requirementsContent []byte,
 	config *types.Configuration) error {
 
 	err := menu.Build(versionsInfo, branches, menuTemplateContent)
@@ -185,32 +177,18 @@ func buildDocumentation(branches []string, versionsInfo types.VersionsInformatio
 		return err
 	}
 
-	err = buildRequirements(versionsInfo, requirementsContent)
+	err = requirements.Build(versionsInfo, requirementsContent)
 	if err != nil {
 		return err
 	}
 
-	baseDockerfile, err := getDockerfile(fallbackDockerfile, versionsInfo.CurrentPath, config.DockerfileName)
+	dockerImageFullName, err := docker.BuildImage(config, fallbackDockerfile, versionsInfo)
 	if err != nil {
-		return err
-	}
-
-	err = ioutil.WriteFile(baseDockerfile.path, baseDockerfile.content, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	dockerImageFullName := buildDockerImageFullName(baseDockerfile.imageName, versionsInfo.Current)
-
-	// Build image
-	output, err := dockerCmd(config.Debug, "build", "--no-cache="+strconv.FormatBool(config.NoCache), "-t", dockerImageFullName, "-f", baseDockerfile.path, versionsInfo.CurrentPath+"/")
-	if err != nil {
-		log.Println(output)
 		return err
 	}
 
 	// Run image
-	output, err = dockerCmd(config.Debug, "run", "--rm", "-v", versionsInfo.CurrentPath+":/mkdocs", dockerImageFullName, "mkdocs", "build")
+	output, err := docker.Exec(config.Debug, "run", "--rm", "-v", versionsInfo.CurrentPath+":/mkdocs", dockerImageFullName, "mkdocs", "build")
 	if err != nil {
 		log.Println(output)
 		return err
@@ -301,145 +279,6 @@ func getMenuFileContent(f string, u string) ([]byte, error) {
 	return content, nil
 }
 
-// checkRequirements return an error if the requirements file is not found in the doc root directory.
-func checkRequirements(docRoot string) error {
-	_, err := os.Stat(filepath.Join(docRoot, "requirements.txt"))
-	return err
-}
-
-func getRequirementsContent(requirementsURL string) ([]byte, error) {
-	var content []byte
-
-	if len(requirementsURL) > 0 {
-		_, err := os.Stat(requirementsURL)
-		if err != nil {
-			content, err = file.Download(requirementsURL)
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to download Requirements file")
-			}
-		} else {
-			content, err = ioutil.ReadFile(requirementsURL)
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to read Requirements file")
-			}
-		}
-	}
-	return content, nil
-}
-
-func buildRequirements(versionsInfo types.VersionsInformation, customContent []byte) error {
-	if len(customContent) > 0 {
-		requirementsPath := filepath.Join(versionsInfo.CurrentPath, "requirements.txt")
-
-		baseContent, err := ioutil.ReadFile(requirementsPath)
-		if err != nil {
-			return err
-		}
-
-		reqBase, err := parseRequirements(baseContent)
-		if err != nil {
-			return err
-		}
-
-		reqCustom, err := parseRequirements(customContent)
-		if err != nil {
-			return err
-		}
-
-		// merge
-		for key, value := range reqCustom {
-			reqBase[key] = value
-		}
-
-		file, err := os.Create(requirementsPath)
-		if err != nil {
-			return err
-		}
-		defer safeClose(file.Close)
-
-		for key, value := range reqBase {
-			fmt.Fprintf(file, "%s%s\n", key, value)
-		}
-	}
-	return nil
-}
-
-func parseRequirements(content []byte) (map[string]string, error) {
-	exp := regexp.MustCompile(`([\w-_]+)([=|>|<].+)`)
-
-	result := make(map[string]string)
-
-	lines := strings.Split(string(content), "\n")
-	for _, line := range lines {
-		if len(line) > 0 {
-			submatch := exp.FindStringSubmatch(line)
-			if len(submatch) != 3 {
-				return nil, errors.Errorf("invalid line format: %s", line)
-			}
-
-			result[submatch[1]] = submatch[2]
-		}
-	}
-
-	return result, nil
-}
-
-// buildDockerImageFullName returns the full docker image name, in the form image:tag.
-// Please note that normalization is applied to avoid forbidden characters
-func buildDockerImageFullName(imageName string, tagName string) string {
-	r := strings.NewReplacer(":", "-", "/", "-")
-	return r.Replace(imageName) + ":" + r.Replace(tagName)
-}
-
-func getDockerfile(fallbackDockerfile dockerfileInformation, workingDirectory string, dockerfileName string) (*dockerfileInformation, error) {
-	if workingDirectory == "" {
-		return nil, errors.New("workingDirectory is undefined")
-	}
-	if _, err := os.Stat(workingDirectory); os.IsNotExist(err) {
-		return nil, err
-	}
-
-	searchPaths := []string{
-		filepath.Join(workingDirectory, dockerfileName),
-		filepath.Join(workingDirectory, "docs", dockerfileName),
-	}
-
-	for _, searchPath := range searchPaths {
-		if _, err := os.Stat(searchPath); !os.IsNotExist(err) {
-			log.Printf("Found Dockerfile for building documentation in %s.", searchPath)
-
-			var dockerFileContent []byte
-			dockerFileContent, err = ioutil.ReadFile(searchPath)
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to get dockerfile file content.")
-			}
-			baseDockerfile := dockerfileInformation{
-				name:      dockerfileName,
-				path:      searchPath,
-				imageName: fallbackDockerfile.imageName,
-				content:   dockerFileContent,
-			}
-			return &baseDockerfile, nil
-		}
-	}
-
-	log.Printf("Using fallback Dockerfile, written into %s", fallbackDockerfile.path)
-	return &fallbackDockerfile, nil
-}
-
-func dockerCmd(debug bool, args ...string) (string, error) {
-	name := "docker"
-
-	if debug {
-		log.Println(name, strings.Join(args, " "))
-	}
-
-	cmd := exec.Command(name, args...)
-	output, err := cmd.CombinedOutput()
-
-	return string(output), err
-}
-
 func cleanAll(workDir string, debug bool) error {
 	err := os.RemoveAll(workDir)
 	if err != nil {
@@ -465,11 +304,4 @@ func createDirectory(directoryPath string) error {
 	}
 
 	return os.MkdirAll(directoryPath, os.ModePerm)
-}
-
-func safeClose(fn func() error) {
-	err := fn()
-	if err != nil {
-		log.Println(err)
-	}
 }
